+++ CHANGELOG.md
@@ -6,3 +6,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+- Added subcommands to listen to PlanktoScope state updates via the MQTT API
+- Added subcommand to start a segmentation routine and wait until it finishes (by default) or until it starts (by command-line flags) or not wait at all (by command-line flags)